FEAT/OrderService 쿠폰 서비스 API 연동 (쿠폰 검증 및 사용/취소 처리) (#177)
* refactor: TimeDeal 연동 요청 DTO의 필드 타입 및 접근 제어자 수정

* feat: 쿠폰 서비스 API 연동 (쿠폰 검증 및 사용/취소 처리)

+++ order-service/src/main/java/com/palja/order_service/application/dto/response/TimeDealRes.java
@@ -20,6 +20,6 @@ public class TimeDealRes {
     private final LocalDateTime endAt;
     private final BigDecimal timeDealPrice;
     private final int discountRate;
-    private final int timeDealStockQuantity;
+    private final Long timeDealStockQuantity;
     private final String status;
 }
